requirements.txt
appdirs==1.4.3
chardet==3.0.4
dateutils==0.6.6
fake-factory==0.5.7
fs==2.1.2
httplib2==0.12.0
incf.countryutils==1.0
incf.dai==0.7
lxml==4.2.5
odict==1.6.2
pkg-resources==0.0.0
psutil==5.4.8
Pygments==2.3.1
pyparsing==2.1.1
python-dateutil==2.7.5
python-slugify==1.2.6
pytz==2018.7
repoze.lru==0.7
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.2.15
typing==3.6.6
ua-parser==0.8.0
Unidecode==1.0.23
user-agents==1.1.0
